// apps/web-svelte/src/lib/api/generated/model/listEntityShelfParams.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export type ListEntityShelfParams = {
kind?: string;
query?: string;
cursor?: string;
hideNsfw?: boolean;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
limit?: number | string;
referencedBy?: string;
relationshipCode?: string;
sort?: string;
sortDirection?: string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
seed?: number | string;
favorite?: boolean;
organized?: boolean;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
ratingMin?: number | string;
/**
 * @pattern ^-?(?:0|[1-9]\d*)$
 */
ratingMax?: number | string;
unrated?: boolean;
status?: string;
bookType?: string;
bookFormat?: string;
nsfw?: boolean;
hasFile?: boolean;
engaged?: boolean;
orphaned?: boolean;
wanted?: boolean;
acquisitionStatus?: string;
};